+++ src/templates/import.js
@@ -1,241 +1,383 @@
 'use strict';
 
-var async = require('async');
-var _ = require('lodash');
-var util = require('../util/util');
-var pjson = require('../../package.json');
-var semver = require('semver');
-var debug = {
-  template: require('debug')('formio:template:template'),
-  items: require('debug')('formio:template:items'),
-  _install: require('debug')('formio:template:_install'),
-  translateSchema: require('debug')('formio:template:translateSchema'),
-  final: require('debug')('formio:template:final'),
-  postResourceInstall: require('debug')('formio:template:postResourceInstall')
+let async = require(`async`);
+let _ = require(`lodash`);
+let util = require(`../util/util`);
+let debug = {
+  template: require(`debug`)(`formio:template:template`),
+  items: require(`debug`)(`formio:template:items`),
+  install: require(`debug`)(`formio:template:install`),
+  updateSchema: require(`debug`)(`formio:template:updateSchema`),
+  final: require(`debug`)(`formio:template:final`),
+  cleanUp: require(`debug`)(`formio:template:cleanUp`)
 };
 
 /**
  * Perform an installation of a specified template.
  *
- * @param formio
- *   The formio object.
+ * @param {Object} router
+ *   The express router object.
  */
-module.exports = function(formio) {
-  var hook = require('../util/hook')(formio);
+module.exports = (router) => {
+  let formio = router.formio;
+  let hook = require(`../util/hook`)(formio);
 
-  // Provide a default alter method.
-  var _alter = function(item, done) {
-    done(null, item);
-  };
+  /**
+   * A base alter used during import, if one wasn`t supplied for the entity.
+   *
+   * @param {Object} item
+   *   The entity being altered.
+   *
+   * @param {Function} done
+   *   The callback function to invoke, with the entity.
+   */
+  let baseAlter = (item, template, done) => done(null, item);
 
-  // Assign the role ids.
-  var assignRoles = function(template, perms) {
-    if (!template.hasOwnProperty('roles')) {
-      return;
+  /**
+   * Converts an entities role id (machineName) to bson id.
+   *
+   * @param {Object} template
+   *   The project memoization of imported entities.
+   * @param {Object} entity
+   *   The role object to convert.
+   *
+   * @returns {boolean}
+   *   Whether or not the conversion was successful.
+   */
+  let roleMachineNameToId = (template, entity) => {
+    if (!entity || !template.hasOwnProperty(`roles`)) {
+      return false;
     }
-    _.each(perms, function(access) {
-      _.each(access.roles, function(role, i) {
-        if (template.roles.hasOwnProperty(role)) {
+
+    if (!(entity instanceof Array)) {
+      if (entity.hasOwnProperty(`role`) && template.roles.hasOwnProperty(entity.role)) {
+        entity.role = template.roles[entity.role]._id.toString();
+        return true;
+      }
+
+      return false;
+    }
+
+    let changes = false;
+
+    // Used for permissions arrays.
+    _.each(entity, (access) => {
+      _.each(access.roles, (role, i) => {
+        if (template.roles.hasOwnProperty(role) && template.roles[role]._id) {
           access.roles[i] = template.roles[role]._id.toString();
+          changes = true;
+        }
+        else {
+          // Remove any unknown roles, they should always be known at this point of the import.
+          delete access.roles[i];
         }
       });
+
+      // Filter any unknown roles from the pruning process.
+      access.roles = _.filter(access.roles);
     });
+
+    return changes;
   };
 
-  // Assign the role to an entity.
-  var assignRole = function(template, entity) {
-    if (!entity) {
+  /**
+   * Converts an entities form id (machineName) to a bson id.
+   *
+   * @param {Object} template
+   *   The project memoization of imported entities.
+   * @param {Object} entity
+   *   The entity object to convert.
+   *
+   * @returns {boolean}
+   *   Whether or not the conversion was successful.
+   */
+  let formMachineNameToId = (template, entity) => {
+    if (!entity.hasOwnProperty(`form`)) {
       return false;
     }
-    if (!template.hasOwnProperty('roles')) {
-      return false;
-    }
-    if (entity.hasOwnProperty('role') && template.roles.hasOwnProperty(entity.role)) {
-      entity.role = template.roles[entity.role]._id.toString();
+    if (template.hasOwnProperty(`forms`) && template.forms.hasOwnProperty(entity.form)) {
+      entity.form = template.forms[entity.form]._id.toString();
       return true;
     }
+    if (template.hasOwnProperty(`resources`) && template.resources.hasOwnProperty(entity.form)) {
+      entity.form = template.resources[entity.form]._id.toString();
+      return true;
+    }
+
     return false;
   };
 
-  // Assign form.
-  var assignForm = function(template, entity) {
-    if (entity.hasOwnProperty('form')) {
-      if (template.hasOwnProperty('forms') && template.forms.hasOwnProperty(entity.form)) {
-        entity.form = template.forms[entity.form]._id.toString();
-        return true;
-      }
-      if (template.hasOwnProperty('resources') && template.resources.hasOwnProperty(entity.form)) {
-        entity.form = template.resources[entity.form]._id.toString();
-        return true;
-      }
-    }
-    return false;
-  };
-
-  // Assign resources.
-  var assignResources = function(template, entity) {
-    if (!entity || !entity.resources || !template.hasOwnProperty('resources')) {
+  /**
+   * Converts an entities resource id (machineName) to a bson id.
+   *
+   * @param {Object} template
+   *   The project memoization of imported entities.
+   * @param {Object} entity
+   *   The entity object to convert.
+   *
+   * @returns {boolean}
+   *   Whether or not the conversion was successful.
+   */
+  let resourceMachineNameToId = (template, entity) => {
+    // Check the template and entity for resource and resources definitions.
+    if (!entity || (!(entity.resource || entity.resources) || !template.hasOwnProperty('resources'))) {
       return false;
     }
-    _.each(entity.resources, function(resource, index) {
-      if (template.resources.hasOwnProperty(resource)) {
-        entity.resources[index] = template.resources[resource]._id.toString();
-      }
-    });
-  };
 
-  // Assign resource.
-  var assignResource = function(template, entity) {
-    if (!entity || !entity.resource || !template.hasOwnProperty('resources')) {
-      return false;
+    let changes = false;
+
+    // Attempt to update resources if present.
+    if (entity.resources && template.hasOwnProperty(`resources`)) {
+      _.each(entity.resources, (resource, index) => {
+        if (template.resources.hasOwnProperty(resource)) {
+          entity.resources[index] = template.resources[resource]._id.toString();
+          changes = true;
+        }
+      });
     }
-    if (template.resources.hasOwnProperty(entity.resource)) {
+
+    // Attempt to update a single resource if present.
+    if (entity.resource && template.hasOwnProperty(`resources`) && template.resources[entity.resource]) {
       entity.resource = template.resources[entity.resource]._id.toString();
-      return true;
+      changes = true;
     }
-    return false;
+
+    return changes;
   };
 
-  // Assign resources within a form.
-  var assignComponent = function(template, components) {
-    var changed = false;
-    util.eachComponent(components, function(component) {
-      if ((component.type === 'resource') && assignResource(template, component)) {
+  /**
+   * Converts any resource id (machineName) references in the form components.
+   *
+   * @param {Object} template
+   *   The project memoization of imported entities.
+   * @param {Array} components
+   *   The form components to scan.
+   *
+   * @returns {boolean}
+   *   Whether or not the any changes were made.
+   */
+  let componentMachineNameToId = (template, components) => {
+    let changed = false;
+    util.eachComponent(components, (component) => {
+      // Update resource machineNames for resource components.
+      if ((component.type === `resource`) && resourceMachineNameToId(template, component)) {
         changed = true;
       }
 
+      // Update resource machineNames for select components with the resource data type.
       if (
-        (component.type === 'select') &&
-        (component.dataSrc === 'resource') &&
-        assignResource(template, component.data)
+        (component.type === `select`) &&
+        (component.dataSrc === `resource`) &&
+        resourceMachineNameToId(template, component.data)
       ) {
-        hook.alter('importComponent', template, components, component.data);
+        hook.alter(`importComponent`, template, component.data);
         changed = true;
       }
 
-      // Allow importing of compoennts.
-      if (hook.alter('importComponent', template, components, component)) {
+      // Allow importing of components.
+      if (hook.alter(`importComponent`, template, component)) {
         changed = true;
       }
     });
+
     return changed;
   };
 
   /**
-   * Methods to fill in all necessary ID's.
-   * @type {{role}}
+   * The list of installable entities. Each entity must have a model and a transform.
+   *
+   * Note: Entities may define a cleanUp function which is executed after the transform is complete.
+   *
+   * @type {*}
    */
-  var parse = {
-    role: function(template, role) {
-      return role;
+  let entities = {
+    role: {
+      model: formio.resources.role.model,
+      valid: (roles) => {
+        if (typeof roles === 'object' && !(roles instanceof Array)) {
+          return true;
+        }
+
+        return false;
+      },
+      transform: (template, role) => role
     },
-    resource: function(template, form) {
-      assignRoles(template, form.submissionAccess);
-      assignRoles(template, form.access);
-      return form;
+    resource: {
+      model: formio.resources.form.model,
+      valid: (resources) => {
+        if (typeof resources === 'object' && !(resources instanceof Array)) {
+          return true;
+        }
+
+        return false;
+      },
+      transform: (template, resource) => {
+        roleMachineNameToId(template, resource.submissionAccess);
+        roleMachineNameToId(template, resource.access);
+        return resource;
+      },
+      cleanUp: (template, resources, done) => {
+        let model = formio.resources.form.model;
+
+        async.forEachOf(resources, (resource, machineName, next) => {
+          if (!componentMachineNameToId(template, resource.components)) {
+            return next();
+          }
+
+          debug.cleanUp(`Need to update resource component _ids for`, machineName);
+          model.findOneAndUpdate(
+            {_id: resource._id, deleted: {$eq: null}},
+            {components: resource.components},
+            {new: true},
+            (err, doc) => {
+              if (err) {
+                return next(err);
+              }
+
+              resources[machineName] = doc.toObject();
+              debug.cleanUp(`Updated resource component _ids for`, machineName);
+              next();
+            }
+          );
+        }, done);
+      }
     },
-    form: function(template, form) {
-      assignRoles(template, form.submissionAccess);
-      assignRoles(template, form.access);
-      assignComponent(template, form.components);
-      return form;
+    form: {
+      model: formio.resources.form.model,
+      valid: (forms) => {
+        if (typeof forms === 'object' && !(forms instanceof Array)) {
+          return true;
+        }
+
+        return false;
+      },
+      transform: (template, form) => {
+        roleMachineNameToId(template, form.submissionAccess);
+        roleMachineNameToId(template, form.access);
+        componentMachineNameToId(template, form.components);
+        return form;
+      }
     },
-    action: function(template, action) {
-      assignForm(template, action);
-      assignResource(template, action.settings);
-      assignResources(template, action.settings);
-      assignRole(template, action.settings);
-      return action;
-    },
-    submission: function(template, submission) {
-      assignForm(template, submission);
-      return submission;
+    action: {
+      model: formio.actions.model,
+      valid: (actions) => {
+        if (typeof actions === 'object' && !(actions instanceof Array)) {
+          return true;
+        }
+
+        return false;
+      },
+      transform: (template, action) => {
+        resourceMachineNameToId(template, action.settings);
+        roleMachineNameToId(template, action.settings);
+
+        // If no changes were made, the form was invalid and we can't insert the action.
+        if (formMachineNameToId(template, action) === false) {
+          return undefined;
+        }
+
+        return action;
+      }
     }
   };
 
-  var postResourceInstall = function(model, template, items, done) {
-    async.forEachOf(items, function(item, name, itemDone) {
-      if (!assignComponent(template, item.components)) {
-        return itemDone();
-      }
+  /**
+   * Installs the given entity.
+   *
+   * @param {Object} entity
+   *   An entity defined in the models object.
+   *
+   * @returns {function()}
+   *   An invokable function to install the entity with callbacks.
+   */
+  let install = (entity) => {
+    let model = entity.model;
+    let valid = entity.valid;
+    let transform = entity.transform;
+    let cleanUp = entity.cleanUp;
 
-      debug.postResourceInstall('Need to update resource component _ids for', name);
-
-      model.findOneAndUpdate(
-        {_id: item._id, deleted: {$eq: null}},
-        {components: item.components},
-        {new: true},
-        function(err, doc) {
-          if (err) {
-            return itemDone(err);
-          }
-          items[name] = doc.toObject();
-          debug.postResourceInstall('Updated resource component _ids for', name);
-          itemDone();
-        }
-      );
-    }, done);
-  };
-
-  // Install a model with a parse method.
-  var _install = function(model, _parse, _postInstall) {
-    return function(template, items, alter, done) {
-      if (!items || _.isEmpty(items)) {
-        return done();
-      }
-
+    return (template, items, alter, done) => {
       // Normalize arguments.
       if (!done) {
         done = alter;
         alter = null;
       }
 
+      // If no items were given for the install, skip this model.
+      if (!items || _.isEmpty(items)) {
+        debug.items(`No items given to install`);
+        return done();
+      }
+
+      alter = alter || baseAlter;
       debug.items(items);
-      alter = alter || _alter;
-      async.forEachOfSeries(items, function(item, name, itemDone) {
-        var document = _parse ? _parse(template, item) : item;
-        document.machineName = name;
-        alter(document, function(err, document) {
+
+      // If the given items don't have a valid structure for this entity, skip the import.
+      if (valid && !valid(items)) {
+        debug.install(`The given items were not valid: ${JSON.stringify(items)}`);
+        return done();
+      }
+
+      async.forEachOfSeries(items, (item, machineName, next) => {
+        let document = transform
+          ? transform(template, item)
+          : item;
+
+        // If no document was provided before the alter, skip the insertion.
+        if (!document) {
+          debug.items(`Skipping item ${item}`);
+          return next();
+        }
+
+        // Set the document machineName using the import value.
+        document.machineName = machineName;
+        alter(document, template, (err, document) => {
           if (err) {
-            return itemDone(err);
+            return next(err);
+          }
+          // If no document was provided after the alter, skip the insertion.
+          if (!document) {
+            debug.install(`No document was given to install after the alter ${item} (${machineName})`);
+            return next();
           }
 
-          debug._install(document);
-          model.findOne({machineName: document.machineName, deleted: {$eq: null}}, function(err, doc) {
+          debug.install(document);
+          model.findOne({machineName: document.machineName, deleted: {$eq: null}}, (err, doc) => {
             if (err) {
-              debug._install(err);
-              return itemDone(err);
+              debug.install(err);
+              return next(err);
             }
             if (!doc) {
-              debug._install('Existing not found (', document.machineName, ')');
+              debug.install(`Existing not found (${document.machineName})`);
               /* eslint-disable new-cap */
               doc = new model(document);
               /* eslint-enable new-cap */
             }
             else {
-              debug._install('Existing found');
+              debug.install(`Existing found`);
               doc = _.assign(doc, document);
-              debug._install(doc);
+              debug.install(doc);
             }
 
-            doc.save(function(err, result) {
+            doc.save((err, result) => {
               if (err) {
-                debug._install(err);
-                return itemDone(err);
+                debug.install(err.errors || err);
+                return next(err);
               }
 
-              debug._install(result);
-              items[name] = result.toObject();
-              itemDone();
+              debug.install(result);
+              items[machineName] = result.toObject();
+              next();
             });
           });
         });
-      }, function(err) {
+      }, (err) => {
         if (err) {
+          debug.install(err);
           return done(err);
         }
-        if (_postInstall) {
-          return _postInstall(model, template, items, done);
+        if (cleanUp) {
+          return cleanUp(template, items, done);
         }
 
         done();
@@ -244,204 +386,55 @@ module.exports = function(formio) {
   };
 
   /**
-   * Translate a schema from older versions to newer versions.
-   * @param template
-   * @param done
+   * Import the formio template.
+   *
+   * Note: This is all of the core entities, not submission data.
+   *
+   * @param {Object} template
+   *   The parsed JSON template to import.
+   * @param {Object} alter
+   *   A map of alter functions.
+   * @param {Function} done
+   *   The callback function to invoke after the import is complete.
+   *
+   * @returns {*}
    */
-  var translateSchema = function(template, done) {
-    // Skip if the template has a correct version.
-    debug.translateSchema('template.version: ', template.version);
-    debug.translateSchema('pjson.templateVersion: ', pjson.templateVersion);
-    if (template.version && !semver.gt(pjson.templateVersion, template.version)) {
-      debug.translateSchema('Skipping');
-      debug.translateSchema(template);
-      return done();
+  let importTemplate = (template, alter, done) => {
+    if (!done) {
+      done = alter;
+      alter = null;
+    }
+    alter = alter || {};
+    if (!template) {
+      return done(`No template provided.`);
     }
 
-    // Create a pick method.
-    var name = function(name) {
-      return function(value) {
-        return value.name === name;
-      };
-    };
+    if (!template.title) {
+      template.title = 'Export';
+    }
+    if (!template.name) {
+      template.name = 'Export';
+    }
 
-    // Fix all schemas.
-    _.each(['forms', 'resources'], function(type) {
-      _.each(template[type], function(form, formKey) {
-        // If no auth action exists, then add a save submission action.
-        var authAction = _.find(template.actions, {name: 'auth', form: formKey});
-        var noSubmit = _.find(template.actions, {name: 'nosubmit', form: formKey});
-        if (!authAction && !noSubmit) {
-          template.actions[formKey + 'Save'] = {
-            title: 'Save Submission',
-            name: 'save',
-            form: formKey,
-            handler: ['before'],
-            method: ['create', 'update'],
-            priority: 10,
-            settings: {}
-          };
-        }
-
-        util.eachComponent(form.components, function(component) {
-          if (component.validate && component.validate.custom) {
-            _.each(template.resources, function(resource) {
-              component.validate.custom = component.validate.custom.replace(resource.name + '.password', 'password');
-            });
-          }
-          if (component.key && component.key.indexOf('.') !== -1) {
-            component.key = component.key.split('.')[1];
-          }
-        });
-      });
-    });
-
-    // Turn all "auth" actions into the new authentication system.
-    _.each(_.pick(template.actions, name('auth')), function(authAction, key) {
-      delete template.actions[key];
-      var userparts = authAction.settings.username.split('.');
-      if (userparts.length <= 1) {
-        return;
+    debug.items(JSON.stringify(template));
+    async.series(hook.alter(`templateSteps`, [
+      async.apply(install(entities.role), template, template.roles, alter.role),
+      async.apply(install(entities.resource), template, template.resources, alter.form),
+      async.apply(install(entities.form), template, template.forms, alter.form),
+      async.apply(install(entities.action), template, template.actions, alter.action)
+    ], install, template), (err) => {
+      if (err) {
+        debug.template(err);
+        return done(err);
       }
 
-      var resource = userparts[0];
-      var username = userparts[1];
-      var password = authAction.settings.password.split('.')[1];
-
-      // Add the Resource action for new associations.
-      if (authAction.settings.association === 'new') {
-        // Ensure that the underlying resource has a role assignment action.
-        var roleAction = _.find(template.actions, {name: 'role', form: resource});
-        if (!roleAction) {
-          template.actions[resource + 'Role'] = {
-            title: 'Role Assignment',
-            name: 'role',
-            'priority': 1,
-            'handler': ['after'],
-            'method': ['create'],
-            'form': resource,
-            'settings': {
-              'association': 'new',
-              'type': 'add',
-              'role': authAction.settings.role
-            }
-          };
-        }
-
-        var fields = {};
-        fields[username] = username;
-        fields[password] = password;
-        template.actions[key + 'SaveResource'] = {
-          title: 'Save Submission',
-          name: 'save',
-          form: authAction.form,
-          handler: ['before'],
-          method: ['create', 'update'],
-          priority: 11,
-          settings: {
-            resource: resource,
-            fields: fields
-          }
-        };
-      }
-
-      // Add the login action.
-      template.actions[key + 'Login'] = {
-        title: 'Login',
-        name: 'login',
-        form: authAction.form,
-        handler: ['before'],
-        method: ['create'],
-        priority: 2,
-        settings: {
-          resources: [resource],
-          username: username,
-          password: password
-        }
-      };
+      debug.final(JSON.stringify(template));
+      done(null, template);
     });
-
-    // Remove all nosubmit actions.
-    _.each(_.pick(template.actions, name('nosubmit')), function(action, key) {
-      delete template.actions[key];
-    });
-
-    // Convert resource actions into Save Submission actions with resource association.
-    _.each(_.pick(template.actions, name('resource')), function(resourceAction, key) {
-      delete template.actions[key];
-
-      var roleAction = _.find(template.actions, {name: 'role', form: resourceAction.settings.resource});
-      if (!roleAction) {
-        template.actions[resourceAction.settings.resource + 'Role'] = {
-          title: 'Role Assignment',
-          name: 'role',
-          'priority': 1,
-          'handler': ['after'],
-          'method': ['create'],
-          'form': resourceAction.settings.resource,
-          'settings': {
-            'association': 'new',
-            'type': 'add',
-            'role': resourceAction.settings.role
-          }
-        };
-      }
-
-      template.actions[key + 'SaveResource'] = {
-        title: 'Save Submission',
-        name: 'save',
-        form: resourceAction.form,
-        handler: ['before'],
-        method: ['create', 'update'],
-        priority: 11,
-        settings: {
-          resource: resourceAction.settings.resource,
-          fields: resourceAction.settings.fields
-        }
-      };
-    });
-
-    done();
   };
 
-  /**
-   * Return an easy way for someone to install a template.
-   */
   return {
-    createInstall: _install,
-    parse: parse,
-    roles: _install(formio.resources.role.model, parse.role),
-    resources: _install(formio.resources.form.model, parse.resource, postResourceInstall),
-    forms: _install(formio.resources.form.model, parse.form),
-    actions: _install(formio.actions.model, parse.action),
-    submissions: _install(formio.resources.submission.model, parse.submission),
-    template: function(template, alter, done) {
-      if (!done) {
-        done = alter;
-        alter = null;
-      }
-      alter = alter || {};
-      if (!template) {
-        return done('No template provided.');
-      }
-
-      debug.items(JSON.stringify(template));
-      async.series([
-        async.apply(translateSchema, template),
-        async.apply(this.roles, template, template.roles, alter.role),
-        async.apply(this.resources, template, template.resources, alter.form),
-        async.apply(this.forms, template, template.forms, alter.form),
-        async.apply(this.actions, template, template.actions, alter.action),
-        async.apply(this.submissions, template, template.submissions, alter.submission)
-      ], function(err) {
-        if (err) {
-          debug.template(err);
-          return done(err);
-        }
-
-        debug.final(template);
-        done(null, template);
-      });
-    }
+    install,
+    template: importTemplate
   };
 };