+++ src/util/email.js
@@ -394,7 +394,7 @@ module.exports = (formio) => {
 
       // If we don't have a valid transport, don't waste time with nunjucks.
       if (!transporter || typeof transporter.sendMail !== 'function') {
-        console.error(`Could not determine which email transport to use for ${emailType}`);
+        debug.error(`Could not determine which email transport to use for ${emailType}`);
         return next();
       }